Fix `dep` to add only sdk package to `vendor`
This adds a rule for `dep` to prune unused packages. Applied for
buildshiprun, gh-push and  git-event.  Also  removes the already
added to vendor packages for modules listed above.

The context is that only the `sdk`  package is used,  but  `dep`
adds the whole `openfaas/openfaas-cloud` repo to `vendor`.

Tested with `dep ensure` and manually checking the vendor folder

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>